requirements.txt
Django==2.2.6
gunicorn==19.9.0
djangorestframework
